fix: Correct cache TTL units from ms to s
